Add directions list to instruction view
