+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/NavigationContract.java
@@ -11,14 +11,6 @@ public interface NavigationContract {
 
     void setSummaryBehaviorHideable(boolean isHideable);
 
-    void setSummaryOptionsVisibility(boolean isVisible);
-
-    void setSummaryDirectionsVisibility(boolean isVisible);
-
-    boolean isSummaryDirectionsVisible();
-
-    void setSheetShadowVisibility(boolean isVisible);
-
     void setCameraTrackingEnabled(boolean isEnabled);
 
     void resetCameraPosition();
@@ -36,21 +28,5 @@ public interface NavigationContract {
     void finishNavigationView();
 
     void setMuted(boolean isMuted);
-
-    void setCancelBtnClickable(boolean isClickable);
-
-    void animateCancelBtnAlpha(float value);
-
-    void animateExpandArrowRotation(float value);
-
-    void animateInstructionViewAlpha(float value);
-
-    int getBottomSheetHeight();
-
-    int getBottomSheetPeekHeight();
-
-    int[] getMapPadding();
-
-    void setMapPadding(int left, int top, int right, int bottom);
   }
 }
