+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/NavigationView.java
@@ -66,11 +66,6 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
   private SummaryBottomSheet summaryBottomSheet;
   private BottomSheetBehavior summaryBehavior;
   private ImageButton cancelBtn;
-  private ImageButton expandArrow;
-  private View summaryDirections;
-  private View summaryOptions;
-  private View directionsOptionLayout;
-  private View sheetShadow;
   private RecenterButton recenterBtn;
   private FloatingActionButton soundFab;
 
@@ -157,7 +152,7 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
   public void onMapReady(MapboxMap mapboxMap) {
     map = mapboxMap;
     map.setOnScrollListener(this);
-    setMapPadding(0, 0, 0, summaryBehavior.getPeekHeight());
+    map.setPadding(0, 0, 0, summaryBehavior.getPeekHeight());
     ThemeSwitcher.setMapStyle(getContext(), map, new MapboxMap.OnStyleLoadedListener() {
       @Override
       public void onStyleLoaded(String style) {
@@ -198,26 +193,6 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
   }
 
   @Override
-  public void setSummaryOptionsVisibility(boolean isVisible) {
-    summaryOptions.setVisibility(isVisible ? View.VISIBLE : View.GONE);
-  }
-
-  @Override
-  public void setSummaryDirectionsVisibility(boolean isVisible) {
-    summaryDirections.setVisibility(isVisible ? View.VISIBLE : View.GONE);
-  }
-
-  @Override
-  public boolean isSummaryDirectionsVisible() {
-    return summaryDirections.getVisibility() == View.VISIBLE;
-  }
-
-  @Override
-  public void setSheetShadowVisibility(boolean isVisible) {
-    sheetShadow.setVisibility(isVisible ? View.VISIBLE : View.GONE);
-  }
-
-  @Override
   public void setCameraTrackingEnabled(boolean isEnabled) {
     camera.setCameraTrackingLocation(isEnabled);
   }
@@ -250,46 +225,6 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
   @Override
   public void setMuted(boolean isMuted) {
     navigationViewModel.setMuted(isMuted);
-  }
-
-  @Override
-  public void setCancelBtnClickable(boolean isClickable) {
-    cancelBtn.setClickable(isClickable);
-  }
-
-  @Override
-  public void animateCancelBtnAlpha(float value) {
-    cancelBtn.animate().alpha(value).setDuration(0).start();
-  }
-
-  @Override
-  public void animateExpandArrowRotation(float value) {
-    expandArrow.animate().rotation(value).setDuration(0).start();
-  }
-
-  @Override
-  public void animateInstructionViewAlpha(float value) {
-    instructionView.animate().alpha(value).setDuration(0).start();
-  }
-
-  @Override
-  public int getBottomSheetHeight() {
-    return summaryBottomSheet.getHeight();
-  }
-
-  @Override
-  public int getBottomSheetPeekHeight() {
-    return summaryBehavior.getPeekHeight();
-  }
-
-  @Override
-  public int[] getMapPadding() {
-    return map.getPadding();
-  }
-
-  @Override
-  public void setMapPadding(int left, int top, int right, int bottom) {
-    map.setPadding(left, top, right, bottom);
   }
 
   /**
@@ -364,11 +299,6 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
     instructionView = findViewById(R.id.instructionView);
     summaryBottomSheet = findViewById(R.id.summaryBottomSheet);
     cancelBtn = findViewById(R.id.cancelBtn);
-    expandArrow = findViewById(R.id.expandArrow);
-    summaryOptions = findViewById(R.id.summaryOptions);
-    summaryDirections = findViewById(R.id.summaryDirections);
-    directionsOptionLayout = findViewById(R.id.directionsOptionLayout);
-    sheetShadow = findViewById(R.id.sheetShadow);
     recenterBtn = findViewById(R.id.recenterBtn);
     soundFab = findViewById(R.id.soundFab);
   }
@@ -387,22 +317,10 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
    * Sets click listeners to all views that need them.
    */
   private void initClickListeners() {
-    directionsOptionLayout.setOnClickListener(new View.OnClickListener() {
-      @Override
-      public void onClick(View view) {
-        navigationPresenter.onDirectionsOptionClick();
-      }
-    });
     cancelBtn.setOnClickListener(new View.OnClickListener() {
       @Override
       public void onClick(View view) {
         navigationPresenter.onCancelBtnClick();
-      }
-    });
-    expandArrow.setOnClickListener(new View.OnClickListener() {
-      @Override
-      public void onClick(View view) {
-        navigationPresenter.onExpandArrowClick(summaryBehavior.getState());
       }
     });
     recenterBtn.setOnClickListener(new View.OnClickListener() {
@@ -424,28 +342,17 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
    */
   private void initSummaryBottomSheet() {
     summaryBehavior = BottomSheetBehavior.from(summaryBottomSheet);
+    summaryBehavior.setHideable(false);
     summaryBehavior.setBottomSheetCallback(new BottomSheetBehavior.BottomSheetCallback() {
       @Override
       public void onStateChanged(@NonNull View bottomSheet, int newState) {
-        switch (newState) {
-          case BottomSheetBehavior.STATE_EXPANDED:
-            navigationPresenter.onSummaryBottomSheetExpanded();
-            break;
-          case BottomSheetBehavior.STATE_COLLAPSED:
-            navigationPresenter.onSummaryBottomSheetCollapsed();
-            break;
-          case BottomSheetBehavior.STATE_HIDDEN:
-            navigationPresenter.onSummaryBottomSheetHidden();
-            break;
-          default:
-            break;
+        if (newState == BottomSheetBehavior.STATE_HIDDEN) {
+          navigationPresenter.onSummaryBottomSheetHidden();
         }
       }
 
       @Override
       public void onSlide(@NonNull View bottomSheet, float slideOffset) {
-        navigationPresenter.onBottomSheetSlide(slideOffset,
-          sheetShadow.getVisibility() == View.VISIBLE);
       }
     });
   }
@@ -457,8 +364,7 @@ public class NavigationView extends CoordinatorLayout implements OnMapReadyCallb
    * @param bottomSheetState retrieved from savedInstanceState
    */
   private void resetBottomSheetState(int bottomSheetState) {
-    boolean isShowing = bottomSheetState == BottomSheetBehavior.STATE_COLLAPSED
-      || bottomSheetState == BottomSheetBehavior.STATE_EXPANDED;
+    boolean isShowing = bottomSheetState == BottomSheetBehavior.STATE_EXPANDED;
     summaryBehavior.setHideable(!isShowing);
     summaryBehavior.setState(bottomSheetState);
   }
