+++ libandroid-navigation-ui/src/main/java/com/mapbox/services/android/navigation/ui/v5/summary/list/DirectionViewHolder.java
@@ -1,6 +1,5 @@
 package com.mapbox.services.android.navigation.ui.v5.summary.list;
 
-import android.graphics.Typeface;
 import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.ImageView;
@@ -9,8 +8,6 @@ import android.widget.TextView;
 import com.mapbox.services.android.navigation.ui.v5.R;
 
 public class DirectionViewHolder extends RecyclerView.ViewHolder {
-
-  private boolean firstViewHolder;
 
   TextView instructionText;
   TextView distanceText;
@@ -23,17 +20,5 @@ public class DirectionViewHolder extends RecyclerView.ViewHolder {
     distanceText = itemView.findViewById(R.id.distanceText);
     directionIcon = itemView.findViewById(R.id.directionIcon);
     distanceDivider = itemView.findViewById(R.id.distanceDivider);
-  }
-
-  public void updateFirstViewHolder() {
-    if (!firstViewHolder) {
-      instructionText.setTypeface(Typeface.DEFAULT_BOLD);
-      firstViewHolder = true;
-    }
-  }
-
-  void resetViewSizes() {
-    instructionText.setTypeface(Typeface.DEFAULT);
-    firstViewHolder = false;
   }
 }
