+++ libandroid-navigation-ui/src/main/res/layout-land/instruction_view_layout.xml
@@ -1,23 +1,32 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
-    android:id="@+id/instructionViewLayout"
     xmlns:android="http://schemas.android.com/apk/res/android"
+    android:id="@+id/instructionViewLayout"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:animateLayoutChanges="true"
     android:keepScreenOn="true">
 
     <include
+        android:id="@+id/instructionLayout"
         layout="@layout/instruction_layout"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/instructionLayout"/>
+        android:layout_height="wrap_content"/>
 
     <include
-        layout="@layout/sound_layout"
+        android:id="@+id/instructionDirectionLayout"
+        layout="@layout/instruction_directions_layout"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_below="@id/instructionLayout"
-        android:id="@+id/soundLayout"/>
+        android:layout_marginTop="-8dp"
+        android:visibility="visible"/>
+
+    <include
+        android:id="@+id/soundLayout"
+        layout="@layout/sound_layout"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_below="@id/instructionLayout"/>
 
 </RelativeLayout>
