+++ libandroid-navigation-ui/src/main/res/layout-land/navigation_view_layout.xml
@@ -16,12 +16,6 @@
         app:mapbox_uiAttribution="false"
         app:mapbox_uiLogo="false"/>
 
-    <com.mapbox.services.android.navigation.ui.v5.instruction.InstructionView
-        android:id="@+id/instructionView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="invisible"/>
-
     <com.mapbox.services.android.navigation.ui.v5.RecenterButton
         android:id="@+id/recenterBtn"
         android:layout_width="wrap_content"
@@ -34,7 +28,12 @@
         android:id="@+id/summaryBottomSheet"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
-        app:behavior_peekHeight="62dp"/>
+        app:layout_behavior="android.support.design.widget.BottomSheetBehavior"/>
+
+    <com.mapbox.services.android.navigation.ui.v5.instruction.InstructionView
+        android:id="@+id/instructionView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:visibility="invisible"/>
 
 </android.support.design.widget.CoordinatorLayout>
