+++ libandroid-navigation-ui/src/main/res/layout-land/summary_peek_layout.xml
@@ -5,6 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="56dp"
+    android:background="?attr/navigationViewPrimary"
     tools:showIn="@layout/summary_content_layout">
 
     <ProgressBar
@@ -21,11 +22,10 @@
         android:id="@+id/cancelBtnLayout"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:background="?attr/selectableItemBackground"
         android:clickable="true"
         android:focusable="true"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true">
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentRight="true">
 
         <ImageButton
             android:id="@+id/cancelBtn"
@@ -39,36 +39,12 @@
 
     </LinearLayout>
 
-    <LinearLayout
-        android:id="@+id/expandArrowLayout"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:background="?attr/selectableItemBackground"
-        android:clickable="true"
-        android:focusable="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true">
-
-        <ImageButton
-            android:id="@+id/expandArrow"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@android:color/transparent"
-            android:layout_gravity="center"
-            android:padding="16dp"
-            android:tint="?attr/navigationViewIcon"
-            app:srcCompat="@drawable/ic_arrow_up"/>
-
-    </LinearLayout>
-
     <include
         layout="@layout/summary_peek_content_layout"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:layout_centerInParent="true"
-        android:layout_toRightOf="@id/cancelBtnLayout"
-        android:layout_toEndOf="@id/cancelBtnLayout"
-        android:layout_toLeftOf="@id/expandArrowLayout"
-        android:layout_toStartOf="@id/expandArrowLayout"/>
+        android:layout_marginLeft="16dp"
+        android:layout_marginStart="16dp"
+        android:layout_centerInParent="true"/>
 
 </RelativeLayout>
