+++ libandroid-navigation-ui/src/main/res/layout/instruction_directions_layout.xml
@@ -4,8 +4,16 @@
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
+    android:background="?attr/navigationViewPrimary"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
+
+    <View
+        android:id="@+id/rvShadow"
+        android:layout_width="wrap_content"
+        android:layout_height="6dp"
+        android:background="@drawable/gradient_shadow_down"
+        android:visibility="invisible"/>
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rvDirections"
