+++ libandroid-navigation-ui/src/main/res/layout/navigation_view_layout.xml
@@ -3,23 +3,17 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/navigationLayout"
-    android:orientation="vertical"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <com.mapbox.mapboxsdk.maps.MapView
         android:id="@+id/mapView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:mapbox_uiCompass="false"
         app:mapbox_uiAttribution="false"
+        app:mapbox_uiCompass="false"
         app:mapbox_uiLogo="false"/>
-
-    <com.mapbox.services.android.navigation.ui.v5.instruction.InstructionView
-        android:id="@+id/instructionView"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:visibility="invisible"/>
 
     <com.mapbox.services.android.navigation.ui.v5.RecenterButton
         android:id="@+id/recenterBtn"
@@ -33,8 +27,12 @@
         android:id="@+id/summaryBottomSheet"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:layout_behavior="android.support.design.widget.BottomSheetBehavior"
-        app:behavior_hideable="false"
-        app:behavior_peekHeight="78dp"/>
+        app:layout_behavior="android.support.design.widget.BottomSheetBehavior"/>
+
+    <com.mapbox.services.android.navigation.ui.v5.instruction.InstructionView
+        android:id="@+id/instructionView"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:visibility="invisible"/>
 
 </android.support.design.widget.CoordinatorLayout>
