+++ libandroid-navigation-ui/src/main/res/layout/summary_peek_layout.xml
@@ -1,74 +1,52 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout
-    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="72dp"
+    android:background="?attr/navigationViewPrimary"
     tools:showIn="@layout/summary_content_layout">
 
     <ProgressBar
         android:id="@+id/rerouteProgressBar"
         style="@style/Widget.AppCompat.ProgressBar"
-        android:indeterminateTint="?attr/navigationViewAccent"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_centerInParent="true"
         android:indeterminate="true"
-        android:visibility="invisible"
-        android:layout_centerInParent="true"/>
+        android:indeterminateTint="?attr/navigationViewAccent"
+        android:visibility="invisible"/>
 
     <LinearLayout
         android:id="@+id/cancelBtnLayout"
         android:layout_width="wrap_content"
         android:layout_height="match_parent"
-        android:background="?attr/selectableItemBackground"
+        android:layout_alignParentEnd="true"
+        android:layout_alignParentRight="true"
         android:clickable="true"
-        android:focusable="true"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentLeft="true">
+        android:focusable="true">
 
         <ImageButton
             android:id="@+id/cancelBtn"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:background="@android:color/transparent"
             android:layout_gravity="center"
+            android:background="@android:color/transparent"
             android:padding="16dp"
             android:tint="?attr/navigationViewIcon"
             app:srcCompat="@drawable/ic_clear"/>
 
     </LinearLayout>
 
-    <LinearLayout
-        android:id="@+id/expandArrowLayout"
-        android:layout_width="wrap_content"
-        android:layout_height="match_parent"
-        android:background="?attr/selectableItemBackground"
-        android:clickable="true"
-        android:focusable="true"
-        android:layout_alignParentEnd="true"
-        android:layout_alignParentRight="true">
-
-        <ImageButton
-            android:id="@+id/expandArrow"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:background="@android:color/transparent"
-            android:layout_gravity="center"
-            android:padding="16dp"
-            android:tint="?attr/navigationViewIcon"
-            app:srcCompat="@drawable/ic_arrow_up"/>
-
-    </LinearLayout>
-
     <include
         layout="@layout/summary_peek_content_layout"
-        android:layout_width="wrap_content"
+        android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_centerInParent="true"
-        android:layout_toRightOf="@id/cancelBtnLayout"
-        android:layout_toEndOf="@id/cancelBtnLayout"
-        android:layout_toLeftOf="@id/expandArrowLayout"
-        android:layout_toStartOf="@id/expandArrowLayout"/>
+        android:layout_marginLeft="16dp"
+        android:layout_marginStart="16dp"
+        android:layout_toLeftOf="@id/cancelBtnLayout"
+        android:layout_toStartOf="@id/cancelBtnLayout"/>
 
 </RelativeLayout>
